TheWatcher/requirements.txt
discord.py==2.3.2
Flask==3.0.3
python-dotenv==1.0.1
scapy==2.5.0
